pyproject.toml
[project]
name = "vaultmind"
version = "0.1.2"
description = "Your personal AI-powered second brain. Feed it anything. Find everything."
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
authors = [{ name = "Rajya Vardhan Singh", email = "imrajyavardhan12@gmail.com" }]
keywords = ["obsidian", "second-brain", "ai", "cli", "knowledge-management"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Text Processing :: Markup :: Markdown",
]
dependencies = [
    "typer>=0.15.0",
    "rich>=13.0.0",
    "textual>=0.70.0",
    "httpx>=0.28.0",
    "trafilatura>=2.0.0",
    "anthropic>=0.52.0",
    "openai>=1.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "tenacity>=9.0.0",
    "structlog>=24.0.0",
    "pyyaml>=6.0.0",
]

[dependency-groups]
dev = [
    "mypy>=1.15.0",
    "pytest>=9.0.0",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.8.0",
]

[project.scripts]
vm = "vaultmind.main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/vaultmind"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
